backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pytesseract
pillow
reportlab
psycopg2-binary
httpx
python-multipart
